fix: unify eslint configs
